Add docs, tests, demo app, and release-prep code

+++ cached_flow/src/main/java/ru/dapadz/cachedflow/cache/keys/LongKey.kt
@@ -11,17 +11,17 @@ import kotlin.reflect.KClass
  * @param name key name
  * @see LongKey
  */
-fun longCacheKey(name: String):LongKey = LongKey(name)
+fun longCacheKey(name: String): LongKey = LongKey(name)
 
 /**
- * A class for saving an item of type [Boolean] to the cache.
+ * A class for saving an item of type [Long] to the cache.
  *
  * @param name key name
  * @see Key
- * @see booleanCacheKey
+ * @see longCacheKey
  */
-class LongKey(name: String): Key<Long>(name) {
-    override fun isTypeOf(valueClass: KClass<*>): Boolean = valueClass == Int::class
+class LongKey(name: String) : Key<Long>(name) {
+    override fun isTypeOf(valueClass: KClass<*>): Boolean = valueClass == Long::class
     override suspend fun getFromStore(store: Store): Flow<Long?> = store.get(StoreKey(name, Long::class))
     override suspend fun saveToStore(item: Long, store: Store) = store.save(StoreKey(name, Long::class), item)
 }
